Refactor: Adding wast line number to AssertReturn for easier debugging; Moving WastCommand sub-classes inside WastCommand
