--- wasp/src/test/java/rrampage/wasp/data/WastAction.java
@@ -1,4 +1,0 @@
-package rrampage.wasp.data;
-
-public record WastAction(String type, String field, WastValue[] args) {
-}
